fix text error message not assembled
