+++ sample/modules/shared/src/commonMain/mock/assets/files/backend/v1/resource/lobby/page-language.json
@@ -1,9 +1,4 @@
 {
-  "setting": {
-    "time-to-live": {
-      "strategy": "single-use"
-    }
-  },
   "root": {
     "setting": {
       "navigation": {
