+++ tuucho/TODO.md
@@ -1,14 +1,13 @@
 ### Todo:
 
-- from error message
-- popup ? selector navigation seem to not work
+- popup ? selector navigation seem to not work ->
+  because selector can't work a cause de NavigationMaterialCacheRepository,
+  un seul choix est gardé en mémoire ou alors routes is changing.
+
+- fix merge style ... previous / next, button
 
 - update test 2e2
 - logic multi level sub resolution + remove hack to list FIrst
-
-
-- NavigationRepositoryProtocol.MaterialCache -> better than class, useless to build
-  an object with lot al suspendlazy class when not all are used...
 
 - screen context better way to use it ?
 - projection input dataclass ?
